refactor(selectors): transition to minimal state selectors to reduce dependencies
- Replaced imports from metamask-state-basic with metamask-state-minimal in multiple files to streamline selector usage.
- Introduced new selectors in metamask-account-selectors.ts to decouple from the main metamask duck, enhancing modularity.
- Added metamask-state-minimal.ts to provide minimal state selectors, breaking circular dependencies in the codebase.
