+++ ui/selectors/metamask-state-minimal.ts
